thumbs/pyproject.toml
[project]
name = "thumbs"
version = "0.0.0"
description = "Lambda function that produces thumbnails from raw leaflets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==4.2.20",
    "pillow-heif>=0.21.0",
    "pillow==11.1.0",
    "sentry-sdk==2.20.0",
    "sorl-thumbnail==12.11.0",
]

[tool.uv]
package = false
